+++ CUDA_CDEM/Domain.cpp
@@ -8,7 +8,7 @@
 #include "aux_functions.h"
 #include "time_measuring.h"
 
-Domain::Domain(int _nelems, int _nnodes, Node * _nodes, Element * _elements, double c_n, double c_s)
+Domain::Domain(int _nelems, int _nnodes, Node * _nodes, Element * _elements, float c_n, float c_s)
 {
 	nelems = _nelems;
 	nnodes = _nnodes;
@@ -30,7 +30,7 @@ Domain::~Domain()
 	delete[] nodes;
 }
 
-void Domain::write_state_to_file(std::string filename, double time)
+void Domain::write_state_to_file(std::string filename, float time)
 {
 	std::ofstream outfile(filename);
 	int i;
@@ -54,11 +54,11 @@ void Domain::write_state_to_file(std::string filename, double time)
 }
 
 // Calculate the force acting on a node as a result of its relative displacement to the neighbor nodes.
-Eigen::Vector2d Domain::get_contact_force(int node_id)
+Eigen::Vector2f Domain::get_contact_force(int node_id)
 {
-	Eigen::Vector2d F(0., 0.);
-	Eigen::Vector2d du_g;
-	Eigen::Matrix2d T;
+	Eigen::Vector2f F(0., 0.);
+	Eigen::Vector2f du_g;
+	Eigen::Matrix2f T;
 	Node& n0 = nodes[node_id-1];
 	int i;
 	for (i = 0; i < 2; i++)
@@ -75,7 +75,7 @@ Eigen::Vector2d Domain::get_contact_force(int node_id)
 }
 
 // Solve the system using the dynamic relaxation method.
-void Domain::solve(double t_load, double t_max, int maxiter)
+void Domain::solve(float t_load, float t_max, int maxiter)
 {	
 	float dt = t_max / maxiter;
 	int i, j;
@@ -89,31 +89,31 @@ void Domain::solve(double t_load, double t_max, int maxiter)
 		}
 	}
 
-	// Eigen matrices will be copied into arrays of doubles.
+	// Eigen matrices will be copied into arrays of floats.
 	// Using the Eigen::Map function defaults in a column by column layout.
-	double * u, *v, *a, *load, *supports, *K, *C, *Mi, *Kc, *n_vects;
+	float * u, *v, *a, *load, *supports, *K, *C, *Mi, *Kc, *n_vects;
 	int * neighbors;
 	int nnodedofs = 2, stiffdim = 8;
 	int vdim = nnodes*nnodedofs;
-	u = new double[vdim];
-	v = new double[vdim];
-	a = new double[vdim];
-	load = new double[vdim];
-	supports = new double[vdim];
-	K = new double[nelems*stiffdim*stiffdim];
-	C = new double[nelems*stiffdim];
-	Mi = new double[nelems*stiffdim];
-	Kc = new double[4];
-	n_vects = new double[4 * nnodes];
+	u = new float[vdim];
+	v = new float[vdim];
+	a = new float[vdim];
+	load = new float[vdim];
+	supports = new float[vdim];
+	K = new float[nelems*stiffdim*stiffdim];
+	C = new float[nelems*stiffdim];
+	Mi = new float[nelems*stiffdim];
+	Kc = new float[4];
+	n_vects = new float[4 * nnodes];
 	neighbors = new int[vdim];
 
 	for (i = 0; i < nnodes; i++)
 	{
 		int of = i*nnodedofs;
-		Eigen::Map<Eigen::VectorXd>(u + of, nodes[i].v_disp.rows(), nodes[i].v_disp.cols()) = nodes[i].v_disp;
-		Eigen::Map<Eigen::VectorXd>(v + of, nodes[i].v_velo.rows(), nodes[i].v_velo.cols()) = nodes[i].v_velo;
-		Eigen::Map<Eigen::VectorXd>(a + of, nodes[i].v_acce.rows(), nodes[i].v_acce.cols()) = nodes[i].v_acce;
-		Eigen::Map<Eigen::VectorXd>(load + of, nodes[i].v_load.rows(), nodes[i].v_load.cols()) = nodes[i].v_load;
+		Eigen::Map<Eigen::VectorXf>(u + of, nodes[i].v_disp.rows(), nodes[i].v_disp.cols()) = nodes[i].v_disp;
+		Eigen::Map<Eigen::VectorXf>(v + of, nodes[i].v_velo.rows(), nodes[i].v_velo.cols()) = nodes[i].v_velo;
+		Eigen::Map<Eigen::VectorXf>(a + of, nodes[i].v_acce.rows(), nodes[i].v_acce.cols()) = nodes[i].v_acce;
+		Eigen::Map<Eigen::VectorXf>(load + of, nodes[i].v_load.rows(), nodes[i].v_load.cols()) = nodes[i].v_load;
 		*(supports++) = nodes[i].supports[0];
 		*(supports++) = nodes[i].supports[1];
 		*(neighbors++) = nodes[i].neighbors[0];
@@ -132,12 +132,12 @@ void Domain::solve(double t_load, double t_max, int maxiter)
 	for (i = 0; i < nelems; i++)
 	{
 		int of1 = i*stiffdim*stiffdim, of2 = i*stiffdim;
-		Eigen::Map<Eigen::MatrixXd>(K + of1, elements[i].K_loc.rows(), elements[i].K_loc.cols()) = elements[i].K_loc;
-		Eigen::Map<Eigen::MatrixXd>(C + of2, elements[i].C_loc.diagonal().rows(), elements[i].C_loc.diagonal().cols()) = elements[i].C_loc.diagonal();
-		Eigen::Map<Eigen::MatrixXd>(Mi + of2, elements[i].M_loc_inv.diagonal().rows(), elements[i].M_loc_inv.diagonal().cols()) = elements[i].M_loc_inv.diagonal();
+		Eigen::Map<Eigen::MatrixXf>(K + of1, elements[i].K_loc.rows(), elements[i].K_loc.cols()) = elements[i].K_loc;
+		Eigen::Map<Eigen::MatrixXf>(C + of2, elements[i].C_loc.diagonal().rows(), elements[i].C_loc.diagonal().cols()) = elements[i].C_loc.diagonal();
+		Eigen::Map<Eigen::MatrixXf>(Mi + of2, elements[i].M_loc_inv.diagonal().rows(), elements[i].M_loc_inv.diagonal().cols()) = elements[i].M_loc_inv.diagonal();
 	}
 
-	Eigen::Map<Eigen::MatrixXd>(Kc, m_contact_stiffness.rows(), m_contact_stiffness.cols()) = m_contact_stiffness;
+	Eigen::Map<Eigen::MatrixXf>(Kc, m_contact_stiffness.rows(), m_contact_stiffness.cols()) = m_contact_stiffness;
 
 	//element_step_with_CUDA(u, v, a, load, supports, neighbors, n_vects, K, C, Mi, Kc, nelems, nnodes, nnodedofs, 
 	//		stiffdim,t_load,t_max,maxiter);
@@ -148,12 +148,12 @@ void Domain::solve(double t_load, double t_max, int maxiter)
 		//}
 }
 
-void Domain::solve(double t_load, double t_max, int maxiter, char* outfile, int output_frequency, int gridDim, int blockDim)
+void Domain::solve(float t_load, float t_max, int maxiter, char* outfile, int output_frequency, int gridDim, int blockDim)
 {
 	float dt = t_max / maxiter;
 	int i, j;
-	double solve_begin_CPU = get_cpu_time();
-	double solve_begin_wall = get_wall_time();
+	float solve_begin_CPU = get_cpu_time();
+	float solve_begin_wall = get_wall_time();
 	for (i = 0; i < nelems; i++)
 	{
 		elements[i].set_matrices();
@@ -164,31 +164,31 @@ void Domain::solve(double t_load, double t_max, int maxiter, char* outfile, int 
 		}
 	}
 
-	// Eigen matrices will be copied into arrays of doubles.
+	// Eigen matrices will be copied into arrays of floats.
 	// Using the Eigen::Map function defaults in a column by column layout.
-	double * u, *v, *a, *load, *supports, *K, *C, *Mi, *Kc, *n_vects;
+	float * u, *v, *a, *load, *supports, *K, *C, *Mi, *Kc, *n_vects;
 	int * neighbors;
 	int nnodedofs = 2, stiffdim = 8;
 	int vdim = nnodes*nnodedofs;
-	u = new double[vdim];
-	v = new double[vdim];
-	a = new double[vdim];
-	load = new double[vdim];
-	supports = new double[vdim];
-	K = new double[nelems*stiffdim*stiffdim];
-	C = new double[nelems*stiffdim];
-	Mi = new double[nelems*stiffdim];
-	Kc = new double[4];
-	n_vects = new double[4 * nnodes];
+	u = new float[vdim];
+	v = new float[vdim];
+	a = new float[vdim];
+	load = new float[vdim];
+	supports = new float[vdim];
+	K = new float[nelems*stiffdim*stiffdim];
+	C = new float[nelems*stiffdim];
+	Mi = new float[nelems*stiffdim];
+	Kc = new float[4];
+	n_vects = new float[4 * nnodes];
 	neighbors = new int[vdim];
 
 	for (i = 0; i < nnodes; i++)
 	{
 		int of = i*nnodedofs;
-		Eigen::Map<Eigen::VectorXd>(u + of, nodes[i].v_disp.rows(), nodes[i].v_disp.cols()) = nodes[i].v_disp;
-		Eigen::Map<Eigen::VectorXd>(v + of, nodes[i].v_velo.rows(), nodes[i].v_velo.cols()) = nodes[i].v_velo;
-		Eigen::Map<Eigen::VectorXd>(a + of, nodes[i].v_acce.rows(), nodes[i].v_acce.cols()) = nodes[i].v_acce;
-		Eigen::Map<Eigen::VectorXd>(load + of, nodes[i].v_load.rows(), nodes[i].v_load.cols()) = nodes[i].v_load;
+		Eigen::Map<Eigen::VectorXf>(u + of, nodes[i].v_disp.rows(), nodes[i].v_disp.cols()) = nodes[i].v_disp;
+		Eigen::Map<Eigen::VectorXf>(v + of, nodes[i].v_velo.rows(), nodes[i].v_velo.cols()) = nodes[i].v_velo;
+		Eigen::Map<Eigen::VectorXf>(a + of, nodes[i].v_acce.rows(), nodes[i].v_acce.cols()) = nodes[i].v_acce;
+		Eigen::Map<Eigen::VectorXf>(load + of, nodes[i].v_load.rows(), nodes[i].v_load.cols()) = nodes[i].v_load;
 		*(supports++) = nodes[i].supports[0];
 		*(supports++) = nodes[i].supports[1];
 		*(neighbors++) = nodes[i].neighbors[0];
@@ -206,18 +206,18 @@ void Domain::solve(double t_load, double t_max, int maxiter, char* outfile, int 
 	for (i = 0; i < nelems; i++)
 	{
 		int of1 = i*stiffdim*stiffdim, of2 = i*stiffdim;
-		Eigen::Map<Eigen::MatrixXd>(K + of1, elements[i].K_loc.rows(), elements[i].K_loc.cols()) = elements[i].K_loc;
-		Eigen::Map<Eigen::MatrixXd>(C + of2, elements[i].C_loc.diagonal().rows(), elements[i].C_loc.diagonal().cols()) = elements[i].C_loc.diagonal();
-		Eigen::Map<Eigen::MatrixXd>(Mi + of2, elements[i].M_loc_inv.diagonal().rows(), elements[i].M_loc_inv.diagonal().cols()) = elements[i].M_loc_inv.diagonal();
+		Eigen::Map<Eigen::MatrixXf>(K + of1, elements[i].K_loc.rows(), elements[i].K_loc.cols()) = elements[i].K_loc;
+		Eigen::Map<Eigen::MatrixXf>(C + of2, elements[i].C_loc.diagonal().rows(), elements[i].C_loc.diagonal().cols()) = elements[i].C_loc.diagonal();
+		Eigen::Map<Eigen::MatrixXf>(Mi + of2, elements[i].M_loc_inv.diagonal().rows(), elements[i].M_loc_inv.diagonal().cols()) = elements[i].M_loc_inv.diagonal();
 	}
 
-	Eigen::Map<Eigen::MatrixXd>(Kc, m_contact_stiffness.rows(), m_contact_stiffness.cols()) = m_contact_stiffness;
+	Eigen::Map<Eigen::MatrixXf>(Kc, m_contact_stiffness.rows(), m_contact_stiffness.cols()) = m_contact_stiffness;
 	std::string num = "00.txt", fl=outfile;
 	write_state_to_file(fl + num, 0);
 	element_step_with_CUDA(u, v, a, load, supports, neighbors, n_vects, K, C, Mi, Kc, nelems, nnodes, nnodedofs,
 		stiffdim, t_load, t_max, maxiter, outfile, output_frequency, gridDim, blockDim);
-	double solve_end_CPU = get_cpu_time();
-	double solve_end_wall = get_wall_time();
+	float solve_end_CPU = get_cpu_time();
+	float solve_end_wall = get_wall_time();
 	std::cout << "CPU time taken: " << solve_end_CPU - solve_begin_CPU << " s" << std::endl;
 	std::cout << "Wall time taken: " << solve_end_wall - solve_begin_wall << " s" << std::endl;
 	//for (j = 0; j < nelems; j++)
@@ -261,10 +261,10 @@ void Domain::load_from_file(std::string filename)
 					nd.ndofs = ndf;
 					nd.neighbors = new int[ndf];
 					nd.supports = new int[ndf];
-					nd.v_load = Eigen::VectorXd::Zero(ndf);
-					nd.v_disp = Eigen::VectorXd::Zero(ndf);
-					nd.v_velo = Eigen::VectorXd::Zero(ndf);
-					nd.v_acce = Eigen::VectorXd::Zero(ndf);
+					nd.v_load = Eigen::VectorXf::Zero(ndf);
+					nd.v_disp = Eigen::VectorXf::Zero(ndf);
+					nd.v_velo = Eigen::VectorXf::Zero(ndf);
+					nd.v_acce = Eigen::VectorXf::Zero(ndf);
 					nd.v_code = new int[ndf];
 					for (i = 0; i < ndf; i++)
 					{
