+++ CUDA_CDEM/Domain.h
@@ -7,20 +7,20 @@ class Element;
 class Domain
 {
 public:
-	Domain(int, int, Node *, Element *, double, double);
+	Domain(int, int, Node *, Element *, float, float);
 	Domain();
 	~Domain();
-	Eigen::Matrix2d m_contact_stiffness;
+	Eigen::Matrix2f m_contact_stiffness;
 	int nelems, nnodes;
 	Element * elements;
 	Node * nodes;
 	// Create a domain structure defined by a text file
 	void load_from_file(std::string filename);
-	void write_state_to_file(std::string filename, double time);
+	void write_state_to_file(std::string filename, float time);
 	// Calculate the force acting on a node as a result of its relative displacement to the neighbor nodes.
-	Eigen::Vector2d get_contact_force(int node_id);
+	Eigen::Vector2f get_contact_force(int node_id);
 	// Solve the system using the dynamic relaxation method.
-	void solve(double t_load, double t_max, int maxiter);
-	void solve(double t_load, double t_max, int maxiter, char * outfile, int output_frequency=1, int gridDim=-1, int blockDim=128);
+	void solve(float t_load, float t_max, int maxiter);
+	void solve(float t_load, float t_max, int maxiter, char * outfile, int output_frequency=1, int gridDim=-1, int blockDim=128);
 };
