+++ CUDA_CDEM/Element.h
@@ -8,21 +8,21 @@ class Element
 public:
 	Element();
 	virtual ~Element();
-	double E;
-	double nu;
-	double density;
-	double thickness;
-	double alfaC;
+	float E;
+	float nu;
+	float density;
+	float thickness;
+	float alfaC;
 	Domain * domain;
 	int nnodes;
 	int * nodes;
 	int stiffness_dim;
-	Eigen::MatrixXd K_loc;
-	Eigen::MatrixXd * B_matrices;
-	Eigen::MatrixXd M_loc;
-	Eigen::MatrixXd M_loc_inv;
-	Eigen::MatrixXd C_loc;
-	double volume;
+	Eigen::MatrixXf K_loc;
+	Eigen::MatrixXf * B_matrices;
+	Eigen::MatrixXf M_loc;
+	Eigen::MatrixXf M_loc_inv;
+	Eigen::MatrixXf C_loc;
+	float volume;
 	// print data
 	void print_self();
 	// Calculate and store local matrices
@@ -30,5 +30,5 @@ public:
 	// Calculate the normal vector of each face and pass it to the adjacent nodes.
 	void calc_normal_vectors();
 	// Perform one iteration of dynamic relaxation. Return the velocity norm.
-	double iterate(float dt, double tau, bool verbose=false);
+	float iterate(float dt, float tau, bool verbose=false);
 };
