+++ CUDA_CDEM/ElementQuadrangleLin.h
@@ -6,13 +6,13 @@ class ElementQuadrangleLin :
 	public Element
 {
 public:
-	ElementQuadrangleLin(double, double, double, double, double, int *,int);
+	ElementQuadrangleLin(float, float, float, float, float, int *, int);
 	ElementQuadrangleLin();
 	~ElementQuadrangleLin();
 	// Calculate local stiffness matrix. Use reduced integration with hourglass stabilization.
 	void set_K_isoparametric();
 	// Calculate stress vector for element gauss points.
 	void set_matrices();
-	Eigen::MatrixXd get_stress();
+	Eigen::MatrixXf get_stress();
 };
