+++ CUDA_CDEM/aux_functions.h
@@ -1,5 +1,5 @@
 #pragma once
-const double coef1 = 0.5;
-const double coef2 = 0.5;
+const float coef1 = 0.5;
+const float coef2 = 0.5;
 
-double load_function(double tau);
+double load_function(float tau);
